Fix for not activating the port menu drop down on the right workspace
I don't feel as though this is an ideal solution. I think a better solution would be to extract the findWorkspace function from Util.py. Put it in a different file/Class. Once that is done then it can be included in UtilConst.py and anything that has a base name of the workspace can have its base replaced with findWorkspace() + <the rest of the name here>