Fixed RegistrationServer.py using old devicebase
RegistrationServer.py was using the old device base still. This file was no longer in the repository but was working because the .pyc file was still there. I fixed this because we shouldn't be relying on a .pyc file for which we don't even have the .py file for anymore.